💚 Move random_ipv4 test usage into iemapp tests
